+++ src/main/java/seedu/fintrack/utils/Parser.java
@@ -1,0 +1,4 @@
+package seedu.fintrack.utils;
+
+public class Parser {
+}
