+++ src/main/java/seedu/fintrack/utils/Storage.java
@@ -1,0 +1,4 @@
+package seedu.fintrack.utils;
+
+public class Storage {
+}
