+++ src/test/java/seedu/fintrack/ExpenseTest.java
@@ -2,6 +2,7 @@ package seedu.fintrack;
 
 import org.junit.jupiter.api.Test;
 import java.util.Date;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ExpenseTest {
 
@@ -9,26 +10,24 @@ class ExpenseTest {
     void getAmount() {
         // Create an Expense with an amount of 1500 cents.
         Expense expense = new Expense(1500, "Food", "Dinner", new Date(1000L));
-        // Use Java assert to verify that getAmount returns 1500.
-        assert expense.getAmount() == 1500 : "Expected amount to be 1500 cents, but got " + expense.getAmount();
+        // Use JUnit assertEquals to verify that getAmount returns 1500.
+        assertEquals(1500, expense.getAmount(), "Expected amount to be 1500 cents");
     }
 
     @Test
     void getCategory() {
         // Create an Expense with the category "Food".
         Expense expense = new Expense(1500, "Food", "Dinner", new Date(1000L));
-        // Use Java assert to verify that getCategory returns "Food".
-        assert "Food".equals(expense.getCategory()) : "Expected category to be 'Food', " +
-                "but got " + expense.getCategory();
+        // Use JUnit assertEquals to verify that getCategory returns "Food".
+        assertEquals("Food", expense.getCategory(), "Expected category to be 'Food'");
     }
 
     @Test
     void getDescription() {
         // Create an Expense with the description "Dinner".
         Expense expense = new Expense(1500, "Food", "Dinner", new Date(1000L));
-        // Use Java assert to verify that getDescription returns "Dinner".
-        assert "Dinner".equals(expense.getDescription()) : "Expected description to be 'Dinner', " +
-                "but got " + expense.getDescription();
+        // Use JUnit assertEquals to verify that getDescription returns "Dinner".
+        assertEquals("Dinner", expense.getDescription(), "Expected description to be 'Dinner'");
     }
 
     @Test
@@ -36,8 +35,7 @@ class ExpenseTest {
         // Create a fixed Date for testing.
         Date testDate = new Date(1000L);
         Expense expense = new Expense(1500, "Food", "Dinner", testDate);
-        // Use Java assert to verify that getDate returns the provided date.
-        assert testDate.equals(expense.getDate()) : "Expected date to match the provided date, " +
-                "but got " + expense.getDate();
+        // Use JUnit assertEquals to verify that getDate returns the provided date.
+        assertEquals(testDate, expense.getDate(), "Expected date to match the provided date");
     }
 }
