+++ src/test/java/seedu/fintrack/FinTrackTest.java
@@ -2,6 +2,8 @@ package seedu.fintrack;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import seedu.fintrack.utils.Ui;
+
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
